Log product information on push+verify events (#4860)
* Log product information on push+verify events

* Renamed to ClientInformation
